added delete confirm

+++ app/scripts/controllers/edit_feedback_controller.js
@@ -5,9 +5,11 @@ actions: {
     this.transitionToRoute('feedback');
    },
   destroy: function () {
+    if (confirm('Are you sure you want to delete your feedback?')) {
     this.get('model').deleteRecord();
     this.get('model').save();
     this.transitionToRoute('feedback');
-   }
+      }
+    }
   }
 });
