+++ app/scripts/controllers/edit_news_controller.js
@@ -5,9 +5,11 @@ actions: {
     this.transitionToRoute('news');
    },
   destroy: function () {
+    if (confirm('Are you sure you want to delete this news item?')) {
     this.get('model').deleteRecord();
     this.get('model').save();
     this.transitionToRoute('news');
-   }
+      }
+    }
   }
 });
